Added more responsive design, logo, favicon, little change in edit post,

+++ src/main/java/com/szync/SzyncPound/controllers/CommentController.java
@@ -3,7 +3,6 @@ package com.szync.SzyncPound.controllers;
 import com.szync.SzyncPound.dto.CommentDto;
 import com.szync.SzyncPound.dto.PostDto;
 import com.szync.SzyncPound.models.Comment;
-import com.szync.SzyncPound.models.Post;
 import com.szync.SzyncPound.models.UserEntity;
 import com.szync.SzyncPound.security.SecurityUtil;
 import com.szync.SzyncPound.service.CommentService;
@@ -12,14 +11,11 @@ import com.szync.SzyncPound.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 
 import java.util.Map;
-import java.util.Optional;
-import java.util.ResourceBundle;
 
 import static com.szync.SzyncPound.mapper.PostMapper.mapToPost;
 
@@ -43,10 +39,9 @@ public class CommentController {
         if (post == null) {
             return ResponseEntity.badRequest().build();
         }
-        UserEntity user = new UserEntity();
         String email = SecurityUtil.getSessionUser();
         if(email != null) {
-            user = userService.findByEmail(email);
+            UserEntity user = userService.findByEmail(email);
             if(user.getId() != commentRequest.getUserId()) {
                 return ResponseEntity.badRequest().build();
             }
@@ -65,15 +60,12 @@ public class CommentController {
     }
 
     @PostMapping("/comment/{commentId}/delete")
-    public ResponseEntity<?> deletePost(@PathVariable("commentId") Long commentId) {
+    public ResponseEntity<Map<String, String>> deletePost(@PathVariable("commentId") Long commentId) {
         UserEntity user = userService.findByEmail(SecurityUtil.getSessionUser());
         Comment comment = commentService.findComment(commentId);
-        if(comment != null) {
-            long postId = comment.getPost().getId();
-            if(user.getId() == comment.getUser().getId() || user.getId() == comment.getPost().getUser().getId()) {
-                commentService.deleteComment(commentId);
-                return ResponseEntity.ok(Map.of("status", "comment deleted"));
-            }
+        if(comment != null && (user.getId() == comment.getUser().getId() || user.getId() == comment.getPost().getUser().getId())) {
+            commentService.deleteComment(commentId);
+            return ResponseEntity.ok(Map.of("status", "comment deleted"));
         }
         return ResponseEntity.badRequest().build();
     }
